Rename config to config-vorago to avoid collisions with default F-Prime config, and move ADC max data/request sizes from AdcTypes.fpp into a configuration FPP
