+++ Va416x0/Types/AdcTypes.fpp
@@ -17,32 +17,19 @@
 module Va416x0 {
 
     @ Status values returned from reading
-    enum AdcSamplerStatus{
+    enum AdcSamplerStatus {
         SUCCESS
         BUSY
     }
 
-    # FIXME:  Long term, I'd like for each deployment to be able to configure it since 
-    # I expect different systems will have different ADC needs
-    # To reduce noise, the Vorago programmer's guide recommends reading samples
-    # multiple times and then averaging the result. So it's reasonable to assume that 
-    # reading a measurement 16x before reading the next will be a common use case. 
-    # The vorago only completes 21.5  16x reads (340) within 1 ms RTI, setting this to 32. 
-    @ Maximum size of the request arrays - value TBR
-    constant ADC_MAX_REQUEST_SIZE = 32
-
-    @ Maximum size of the data arrays - value TBR
-    constant ADC_MAX_DATA_SIZE = 32
-
-    @ Values read from an ADC channel are 12-bits (there is an option to store the 
-    @ channel ID in the top 4 bits, but it's not enabled) 
-    @ If a channel is read multiple times (limit 16), the sum of all reads is 
+    @ Values read from an ADC channel are 12-bits (there is an option to store the
+    @ channel ID in the top 4 bits, but it's not enabled)
+    @ If a channel is read multiple times (limit 16), the sum of all reads is
     @ stored as the value in this array
-    @ Larger than  
     array AdcData = [ADC_MAX_DATA_SIZE] U16
 
-    @ AdcSampler & AdcCollector will convert each entry in this array to a bit-packed 
+    @ AdcSampler & AdcCollector will convert each entry in this array to a bit-packed
     @ struct 4-bytes in size but FPP doesn't support bit packed arrays, so use U32s instead
-    array AdcRequests = [ADC_MAX_REQUEST_SIZE] U32 
+    array AdcRequests = [ADC_MAX_REQUEST_SIZE] U32
 
 }
